+++ apollo-portal/src/main/resources/static/scripts/controller/IndexController.js
@@ -15,21 +15,26 @@
  *
  */
 index_module.controller('IndexController', ['$scope', '$window', '$translate', 'toastr', 'AppUtil', 'AppService',
-    'UserService', 'FavoriteService',
-    IndexController]);
+    'UserService', 'FavoriteService', 'NamespaceService',
+    IndexController]
+)
 
-function IndexController($scope, $window, $translate, toastr, AppUtil, AppService, UserService, FavoriteService) {
+
+function IndexController($scope, $window, $translate, toastr, AppUtil, AppService, UserService, FavoriteService, NamespaceService) {
 
     $scope.userId = '';
+    $scope.whichContent = '1';
 
     $scope.getUserCreatedApps = getUserCreatedApps;
     $scope.getUserFavorites = getUserFavorites;
-
+    $scope.getPublicNamespaces = getPublicNamespaces;
     $scope.goToAppHomePage = goToAppHomePage;
     $scope.goToCreateAppPage = goToCreateAppPage;
     $scope.toggleOperationBtn = toggleOperationBtn;
     $scope.toTop = toTop;
     $scope.deleteFavorite = deleteFavorite;
+    $scope.morePublicNamespace = morePublicNamespace;
+    $scope.changeContent = changeContent;
 
     function initCreateApplicationPermission() {
         AppService.has_create_application_role($scope.userId).then(
@@ -51,6 +56,10 @@ function IndexController($scope, $window, $translate, toastr, AppUtil, AppServic
         $scope.favoritesPage = 0;
         $scope.favorites = [];
         $scope.hasMoreFavorites = true;
+        $scope.publicNamespacePage = 0;
+        $scope.publicNamespaces = [];
+        $scope.hasMorePublicNamespaces = true;
+        $scope.allPublicNamespaces = [];
         $scope.visitedApps = [];
 
         initCreateApplicationPermission();
@@ -59,7 +68,10 @@ function IndexController($scope, $window, $translate, toastr, AppUtil, AppServic
 
         getUserFavorites();
 
+        getPublicNamespaces();
+
         initUserVisitedApps();
+
     });
 
     function getUserCreatedApps() {
@@ -114,8 +126,31 @@ function IndexController($scope, $window, $translate, toastr, AppUtil, AppServic
                             app.favoriteId = favorite.id;
                             $scope.favorites.push(app);
                         });
-
                     });
+            })
+    }
+
+    function getPublicNamespaces() {
+        NamespaceService.find_public_namespaces()
+            .then(function (result) {
+                $scope.allPublicNamespaces = result;
+                morePublicNamespace();
+                var selectResult = [];
+                angular.forEach(result,function (app) {
+                    selectResult.push({
+                        id: app.appId,
+                        text: app.appId + ' / ' + app.name
+                    })
+                });
+                $('#public-name-spaces-search-list').select2({
+                    data: selectResult,
+                });
+                $('#public-name-spaces-search-list').on('select2:select', function () {
+                    var selected = $('#public-name-spaces-search-list').select2('data');
+                    if (selected && selected.length) {
+                        goToAppHomePage(selected[0].id)
+                    }
+                });
             })
     }
 
@@ -180,6 +215,25 @@ function IndexController($scope, $window, $translate, toastr, AppUtil, AppServic
         $scope.hasMoreFavorites = true;
 
         getUserFavorites();
+    }
+
+    function morePublicNamespace() {
+        var rest = $scope.allPublicNamespaces.length - $scope.publicNamespacePage * 10;
+        if (rest <= 10) {
+            for (var i = 0; i < rest; i++) {
+                $scope.publicNamespaces.push($scope.allPublicNamespaces[$scope.publicNamespacePage * 10 + i])
+            }
+            $scope.hasMorePublicNamespaces = false;
+        } else {
+            for (var j = 0; j < 10; j++) {
+                $scope.publicNamespaces.push($scope.allPublicNamespaces[$scope.publicNamespacePage * 10 + j])
+            }
+        }
+        $scope.publicNamespacePage += 1;
+    }
+
+    function changeContent(contentIndex) {
+        $scope.whichContent = contentIndex;
     }
 
 }